# Conflicts:
#	.gitignore
#	README.md
#	lab1/factorial/CMakeLists.txt
#	lab1/factorial/Factorial.cpp
#	lab1/factorial/Factorial.h
#	lab2/CMakeLists.txt
#	lab4/CMakeLists.txt
#	tests/lab2test/CMakeLists.txt
#	tests/lab3test/CCounterTest.cpp
#	tests/lab3test/CMakeLists.txt
#	tests/lab3test/SmartTreeStep4Test.cpp
#	tests/lab4test/CMakeLists.txt
#	tests/lab4test/SimpleJsonTest.cpp
#	tests/lib/gtest-1.8.0/googlemock/msvc/2005/gmock.sln
#	tests/lib/gtest-1.8.0/googlemock/msvc/2005/gmock.vcproj
#	tests/lib/gtest-1.8.0/googlemock/msvc/2005/gmock_config.vsprops
#	tests/lib/gtest-1.8.0/googlemock/msvc/2005/gmock_main.vcproj
#	tests/lib/gtest-1.8.0/googlemock/msvc/2005/gmock_test.vcproj
#	tests/lib/gtest-1.8.0/googlemock/msvc/2010/gmock.sln
#	tests/lib/gtest-1.8.0/googlemock/msvc/2010/gmock.vcxproj
#	tests/lib/gtest-1.8.0/googlemock/msvc/2010/gmock_config.props
#	tests/lib/gtest-1.8.0/googlemock/msvc/2010/gmock_main.vcxproj
#	tests/lib/gtest-1.8.0/googlemock/msvc/2010/gmock_test.vcxproj
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest.cbproj
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest.groupproj
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest_all.cc
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest_link.cc
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest_main.cbproj
#	tests/lib/gtest-1.8.0/googletest/codegear/gtest_unittest.cbproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest-md.sln
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest-md.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest.sln
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_main-md.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_main.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_prod_test-md.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_prod_test.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_unittest-md.vcproj
#	tests/lib/gtest-1.8.0/googletest/msvc/gtest_unittest.vcproj
#	tests/lib/stringutility/StringUtility.h

+++ lab1/factorial/Factorial.h
@@ -5,6 +5,6 @@
 #ifndef JIMP_EXERCISES_FACTORIAL_H
 #define JIMP_EXERCISES_FACTORIAL_H
 
-long long int factorial(long long int value);
+int factorial(int value);
 
 #endif //JIMP_EXERCISES_FACTORIAL_H
